Allow dashes in user urn (#1564)
* Fix: allow user/group urns to contain dashes

CorpUser urns containing dashes are valid entries. When adding that user as an
owner, the MAE job validates the owner's urn using a regex filter that only accepts
alphanumeric characters and underscore (\w). That means any ownership changes where
the user urn contains an underscore are rejected.

This change extends the regex filter to allow dashes in the name. It includes unit
tests that verify the change works for multiple dashes and underscores.

There are other cases to consider:

1. Should any other characters be allowed?
2. Should the filter check the urn starts and ends with alphanumeric characters?

CLOSES: User urn does not handle dashes consistently #1554

BREAKING CHANGE: None. This change relaxes a restriction so existing code is ok.

* Added tests for group members and fixed assertion